# perf(proto): improve `ColStr` writing performance for small strings